refactor: modularize `postRequest`, `highlightCode`, and `renderMarkdown`
