+++ src/utils/renderMarkdown.js
@@ -1,0 +1,7 @@
+export function renderMarkdown() {
+    window.renderMarkdown = function (content) {
+        const md = new markdownit();
+        return md.render(content);
+    };
+}
+
